+++ examples/LVGL/v8/Rotation/lvgl_port_v8.h
@@ -5,6 +5,7 @@
  */
 #pragma once
 
+#include <Arduino.h>
 #include <ESP_Panel_Library.h>
 #include <lvgl.h>
 
@@ -49,7 +50,9 @@
 #define LVGL_PORT_TASK_MIN_DELAY_MS             (2)         // The minimum delay of the LVGL timer task, in milliseconds
 #define LVGL_PORT_TASK_STACK_SIZE               (6 * 1024)  // The stack size of the LVGL timer task, in bytes
 #define LVGL_PORT_TASK_PRIORITY                 (2)         // The priority of the LVGL timer task
-#define LVGL_PORT_TASK_CORE                     (-1)        // The core of the LVGL timer task, `-1` means the don't specify the core
+#define LVGL_PORT_TASK_CORE                     (ARDUINO_RUNNING_CORE)
+                                                            // The core of the LVGL timer task, `-1` means the don't specify the core
+                                                            // Default is the same as the Arduino task
                                                             // This can be set to `1` only if the SoCs support dual-core,
                                                             // otherwise it should be set to `-1` or `0`