+++ src/touch/base/esp_lcd_touch.h
@@ -27,7 +27,7 @@ extern "C" {
 
 #define ESP_LCD_TOUCH_VER_MAJOR    (1)
 #define ESP_LCD_TOUCH_VER_MINOR    (1)
-#define ESP_LCD_TOUCH_VER_PATCH    (1)
+#define ESP_LCD_TOUCH_VER_PATCH    (2)
 
 #define CONFIG_ESP_LCD_TOUCH_MAX_POINTS     (ESP_PANEL_TOUCH_MAX_POINTS)
 #define CONFIG_ESP_LCD_TOUCH_MAX_BUTTONS    (ESP_PANEL_TOUCH_MAX_BUTTONS)
@@ -69,10 +69,12 @@ typedef struct {
 
     /*!< User callback called after get coordinates from touch controller for apply user adjusting */
     void (*process_coordinates)(esp_lcd_touch_handle_t tp, uint16_t *x, uint16_t *y, uint16_t *strength, uint8_t *point_num, uint8_t max_point_num);
-    /*!< User callback called after the touch interrupt occured */
+    /*!< User callback called after the touch interrupt occurred */
     esp_lcd_touch_interrupt_callback_t interrupt_callback;
     /*!< User data passed to callback */
     void *user_data;
+    /*!< User data passed to driver */
+    void *driver_data;
 } esp_lcd_touch_config_t;
 
 typedef struct {
@@ -387,7 +389,7 @@ esp_err_t esp_lcd_touch_get_mirror_y(esp_lcd_touch_handle_t tp, bool *mirror);
 esp_err_t esp_lcd_touch_del(esp_lcd_touch_handle_t tp);
 
 /**
- * @brief Register user callback called after the touch interrupt occured
+ * @brief Register user callback called after the touch interrupt occurred
  *
  * @param tp: Touch handler
  * @param callback: Interrupt callback
@@ -398,7 +400,7 @@ esp_err_t esp_lcd_touch_del(esp_lcd_touch_handle_t tp);
 esp_err_t esp_lcd_touch_register_interrupt_callback(esp_lcd_touch_handle_t tp, esp_lcd_touch_interrupt_callback_t callback);
 
 /**
- * @brief Register user callback called after the touch interrupt occured with user data
+ * @brief Register user callback called after the touch interrupt occurred with user data
  *
  * @param tp: Touch handler
  * @param callback: Interrupt callback