+++ src/touch/base/esp_lcd_touch_gt911.h
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2022-2023 Espressif Systems (Shanghai) CO LTD
+ * SPDX-FileCopyrightText: 2022-2024 Espressif Systems (Shanghai) CO LTD
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -19,7 +19,7 @@ extern "C" {
 
 #define ESP_LCD_TOUCH_GT911_VER_MAJOR    (1)
 #define ESP_LCD_TOUCH_GT911_VER_MINOR    (1)
-#define ESP_LCD_TOUCH_GT911_VER_PATCH    (0)
+#define ESP_LCD_TOUCH_GT911_VER_PATCH    (1)
 
 /**
  * @brief Create a new GT911 touch driver
@@ -46,12 +46,38 @@ esp_err_t esp_lcd_touch_new_i2c_gt911(const esp_lcd_panel_io_handle_t io, const 
 #define ESP_LCD_TOUCH_IO_I2C_GT911_ADDRESS_BACKUP   (0x14)
 
 /**
+ * @brief GT911 Configuration Type
+ *
+ */
+typedef struct {
+    uint8_t dev_addr;  /*!< I2C device address */
+} esp_lcd_touch_io_gt911_config_t;
+
+/**
  * @brief Touch IO configuration structure
  *
  */
 #define ESP_LCD_TOUCH_IO_I2C_GT911_CONFIG()           \
     {                                       \
         .dev_addr = ESP_LCD_TOUCH_IO_I2C_GT911_ADDRESS, \
+        .control_phase_bytes = 1,           \
+        .dc_bit_offset = 0,                 \
+        .lcd_cmd_bits = 16,                 \
+        .flags =                            \
+        {                                   \
+            .disable_control_phase = 1,     \
+        }                                   \
+    }
+
+/**
+ * @brief Touch IO configuration structure with input address
+ *
+ * @param[in] addr I2C address of the touch panel
+ *
+ */
+#define ESP_LCD_TOUCH_IO_I2C_GT911_CONFIG_WITH_ADDR(addr) \
+    {                                       \
+        .dev_addr = addr,                   \
         .control_phase_bytes = 1,           \
         .dc_bit_offset = 0,                 \
         .lcd_cmd_bits = 16,                 \