+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # ChangeLog
 
-## v0.1.5 - 2024-07-01
+## v0.1.5 - 2024-07-09
+
+### Enhancements:
+
+* feat(gt911): support set I2C address by using RST and INT pins
+* feat(lvgl_port): set the lvgl task to run on the same core as the Arduino task by default
+* feat(board): increase the RGB pclk frequency to 26MHz for `ESP32_4848S040C_I_Y_3`
 
 ### Bugfixes:
 
 * fix(panel): init expander host with correct macro (#65)
+* fix(panel): don't reset the LCD if the bus is RGB bus and the `ESP_PANEL_LCD_FLAGS_AUTO_DEL_PANEL_IO` is enabled
+* fix(examples): fix lvgl port rotation issue when enabling avoid tearing by @NecroMancer05
 
 ## v0.1.4 - 2024-06-17